Added Animation Logic to Player
I've added an animator with a bunch of default animations to the game. I've also added a section in the script that tell the animator which animation to play. The animations are currently empty meaning that they don't actually do anything when played, but it should be simple to add the animations in once they are created.